Write custom `Stack` printing code. More TODO, renames, and changes.
Most importantly, we have added a big TODO in `Instructions.fs`
concerning the current issue with the "weak" wrapping stack operations
since they do not work as indented in the VM code currently and will
need to be fixed.

The main program itself will now generate a random bytecode program and
try to run it.

More renames and minor changes.