Remove lambda feature for simplified handling for insufficient arguments
Since the lambda feature was getting a bit too complicated with also the
issue of having the stack not having structural equality which made
things a bit more complicated was not especially helping.

Likewise, the VM architecture has been significantly simplified where
operation will either succeed or fail (usually in the case of
insufficient arguments). A failed operation will then just be converted
to its raw byte value and then pushed on the stack. This solved the
"not enough args" problem with the VM design.

Also in consideration that now operations themselves cannot be pushed
onto the stack, we have added a `LastOperation` field to the `Engine`
record type. This will report back the last operation that the engine
executed (see the `Engine.step` implementation and the record type
itself for more info).

Also from adding the `LastOperation` field, we found out that we were
never clearing the `LastOutput` field. Fixed that bug.

Display the `LastOperation` field in `HiTop.Program` output.

Fixes bug in `HiTop.Program` were we were using an old `engine` value
instead of `engine'`.

The `HiTop.Program` will now generate more "distributed" random programs
by choosing from generating an unencoded byte or an operation
separately.