Implement brainf**k style loops to make the VM Turing complete
This adds brainf**k style while loops by using begin and end markers
for the loops. This still of course needs to be tested.

Remove `Engine.NextReadAddress` since everything relies on the
underlying stream's position anyways for determining which byte will be
read next in the program. This has been replaced with the equivalent
wrapper functions.

Add `hitop_true` and `hitop_false` literals to avoid sprinkling `0uy`
and `1uy` as magic numbers for the equivalents of the boolean values.

Additionally much of `Engine` module backend structure has been altered.