Implement genetic algorithm library. Support loading and unloading prgms
This adds `HiTop.GeneticAlgorithm` which allows for using a genetic
algorithm over `HiTop.VM`. Of course, we still need the actual runner
project to get anywhere with it.

Refactor the `HiTop.VM.Engine` module to at the base, create an `Engine`
that has to program loaded into it until you load one in. The point of
this refactoring is so that the genetic algorithm make essentially
copies of this "empty" engine that already has the instruction set
loaded in but not a program.

Remove `Result` in `HiTop.VM.InstructionSet` and just use an assertion.

Create `HiTop.VM.Stack.empty` as alias to `Stack.create ()`

Move `InstructionSet` to be before `Engine` since we now have references
to it and it kinda makes sense in terms of going to from low level to
high level.

Update `HiTop.VM.Runner` program to reflect changes to the VM
architecture.