Merge raw bytes into `ByteCode`. Refactor how InstructionSets work.
To make this a bit simpler and also bit more versatile, values
themselves are now mapped to bytes which removes the mess of having to
use nested pattern matching. This also removes the need for the check
to see if the byte is registers in the map and now just has to do a
look up `FromByte`.

This updates `CoreTypes` and `InstructionSet`.

To make life also a bit easier, we are kinda cheating but we are making
`Intruction` have structural equality by using `ShortName` as
essentially a key. Of course this only holds up as long as there are no
duplicate instructions or instructions with the same name but different
function code. At least we have a check at the end of the module to make
sure that all the built-in functions have unique names. Of course, we
could always have a master list of instructions that would have to be
added to ensure that all instructions ever created have unique names
anyhow.

This applies to mainly `CoreTypes.BuiltInstructionSet` in how we are
using ByteCode as a key.

Add `Instructions.toString`

Update `VM.Runner` program in accordance with these changes.