Create project `HiTop.GeneticAlgorithm.Runner` to run the genetic algo
With now having the genetic algorithm runner in place, we can now start
testing and trying to optimize the genetic algorithm so that we can try
to achieve our goal.

Right now we have the genetic algorithm running off of a small sample of
text to see if it can manage to match it correctly with a generated
program.

Add fancy copyright header on the genetic algo runner.

Some the the types in `HiTop.GeneticAlgorithm` have been renamed with a
`T` appended to that its easier to use intellsense with the module
functions then trying to go through the tons of functions for a
built-in type we are aliasing.

Move the random byte generator to a new module `VM.Helpers` with
"pregen" to avoid fetching all the operation bytes every single time.

Inline a ton of stuff that was getting called millions of times along
other performance bottlenecks discovered in profiling.

Of course, the stream code needs to be using `async` which was clearly
overlooked so that the fitness function can be parallelized in a thread
safe way.