Write code for dummy fitness eval, part of selection, and stats output
We have started on the part of the code for the evaluation and selection
part of the genetic algorithm. Of course we are still missing the rest
of selection and cross over to have completed the entire loop for the
genetic algorithm.

Right now though, we have the evaluation with the fitness function there
will a dummy fitness function there for now until we re-implement the
virtual machine to actually run the bytecode. Additionally, the same
statistics output has been added so that the user will be able to tell
how the genetic algorithm is doing.

Also move `try_read_file(..)` to `util.cpp`

Still more work to be done...