Make the CUDA code compile and actually run without crashing.
Most of the lambdas have been replaced with regular functions since they
where causing plenty of trouble.

There also was something weird happening with the `SelectionResult`
struct constructors so those have just been moved to
`__host__ __device__` helper functions for creating the structs.

Also moved `SelectionResult` and related things to its own namespace.

`*program.descriptors.end()` is of course one past the end of the vector
which would result in the program blowing up. This of course with the
other `.being()` call has been replace with `.front()` and `.back()`
which is what we wanted in the first place.